Fix implicit resolution
